+++ example/integration_test/chargebee_test.dart
@@ -128,7 +128,7 @@ class ChargebeeTest {
     }
 
     try {
-      final result = await Chargebee.purchaseProduct(product, customer: customer);
+      final result = await Chargebee.purchaseStoreProduct(product, customer: customer);
       debugPrint('purchase result: $result');
       expect(result.status, 'true');
       tester.printToConsole('Product subscribed successfully!');
@@ -147,7 +147,7 @@ class ChargebeeTest {
     }
 
     try {
-      final result = await Chargebee.purchaseProduct(product, customer: customer);
+      final result = await Chargebee.purchaseStoreProduct(product, customer: customer);
       debugPrint('purchase result: $result');
       expect(result.status, 'true');
       tester.printToConsole('Product subscribed successfully!');